extract macros into own file
